+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v2.2.0 - 2022-12-23
+
+### Added
+
+- Iterable and Iterator implementations to `Scru128Generator` to make it work as
+  infinite iterator
+
+### Maintenance
+
+- Update gradlew and compile-time/development dependencies
+
 ## v2.1.2 - 2022-06-11
 
 ### Fixed